+++ lib/objects/animated_object.dart
@@ -15,12 +15,12 @@ class AnimatedObject extends GameComponent {
 
   @override
   void render(Canvas canvas) {
+    super.render(canvas);
     animation?.getSprite().renderFromVector2Rect(
           canvas,
           this.position,
           opacity: opacity,
         );
-    super.render(canvas);
   }
 
   @override