+++ lib/game_interface/interface_component.dart
@@ -55,12 +55,12 @@ class InterfaceComponent extends GameComponent with TapGesture {
 
   @override
   void render(Canvas canvas) {
+    super.render(canvas);
     (selected ? spriteSelected : sprite)?.renderFromVector2Rect(
       canvas,
       this.position,
       opacity: opacity,
     );
-    super.render(canvas);
   }
 
   @override